Save project data in the SCENARIO block instead of an external file
Storing the projects' launch counts in a separate config file means that
they're shared between all the game's .sfs files, so they don't change
when the game loads a different .sfs file.  In particular, when the
player loads an old save, or reverts from flight back to the VAB/SPH,
the launch counts don't reset back to their earlier values.  This can
cause gaps in the numbering sequence, because numbers get "used up" by
launches that effectively never happened (because the player went back
to a save from before the launch).

To prevent that, I've changed the ProjectManagerScenario's OnSave/OnLoad
methods to store the project data in the SCENARIO block in the .sfs file
instead.  That means that each .sfs file has its own set of launch
counts, and loading an old save will restore the counts to what they
were at that time.  In particular, launching, reverting, and launching
again will produce the same mission number both times.

For backward compatibility, if the launch counts are missing from the
.sfs file, they'll be loaded from the old ProjectManager.settings file
as a fallback.  This ensures that players won't lose their launch counts
when they upgrade to this version.  Going forward, the launch counts are
written only to the .sfs file, not to the settings file.

Once the launch counts have been imported into the player's .sfs
file(s), the ProjectManager.settings file is no longer needed and can be
safely deleted (though there's no harm in leaving it there).  New games
won't create the file in the first place.